pre-capstone/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pytesseract==0.3.10
face-recognition==1.3.0
opencv-python==4.9.0.80
Pillow==10.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
cryptography==41.0.5
apscheduler==3.10.4